Split check_cast to a separate file
